Change name new-story-submit and update the submit
